Tester script with some edits
Tester Files available. The script is saves in "Tester.py", and the file we try to process is in "test.ply". I've added the code I currently have for shifting everything to the origin, seems like it's working (as in nothing changes, I'm pretty sure the original tester.ply file is already centered at the origin), but I can do some more verification. Next steps for me is to try to code rotations