Fixed parameters -- May need redo, read comments
Fixed some obvious stuff with things being called with different combinations of parameters across the code.

Lines 32 and 34 in jobrun could be wrong, but at least we went further in testing with this fix.

With these fixed the test_flow stops working at jobrun line 50. Not sure what is going but want to commit at least these for now.